update appdbcontext, transactionparking, transactionparkingexit_input, farescheme

+++ CustomModels/Fare/FareScheme.cs
@@ -3,7 +3,7 @@
     public class FareScheme
     {
         public decimal Fare { get; set; }
-        public decimal UnitHour { get; set; }
+        public int UnitHour { get; set; }
         public int MaxScheme { get; set; } //0 for unlimited
     }
 }
